some bugs

+++ src/main/java/rtree/DivisionPerformer.java
@@ -4,5 +4,5 @@ import java.util.Set;
 
 public interface DivisionPerformer {
 
-  Set<SpatialKey> divide(double numOfSubNodes);
+  Set<SpatialKey> divide(int numOfSubNodes);
 }
